# openapi.json
{"openapi":"3.1.0","info":{"title":"BaZi True Solar API","description":"Compute BaZi (八字) using true solar time (真太阳时) from user inputs.","version":"1.0.0"},"paths":{"/bazi/compute":{"post":{"summary":"Compute","operationId":"compute_bazi_compute_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaziRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaziResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BaziRequest":{"properties":{"calendar":{"type":"string","const":"gregorian","title":"Calendar","default":"gregorian"},"year":{"type":"integer","title":"Year"},"month":{"type":"integer","title":"Month"},"day":{"type":"integer","title":"Day"},"time":{"type":"string","title":"Time","description":"HH or HH:MM or HH:MM:SS (also accepts HH.MM)"},"gender":{"type":"string","enum":["male","female"],"title":"Gender","default":"male"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"},"lon":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lon","description":"Longitude (east positive). Preferred for accuracy."},"lat":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lat"},"tz":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tz","description":"IANA timezone like Asia/Shanghai. If omitted and country==China -> Asia/Shanghai."},"use_dst":{"type":"boolean","title":"Use Dst","default":false}},"type":"object","required":["year","month","day","time"],"title":"BaziRequest"},"BaziResponse":{"properties":{"input":{"additionalProperties":true,"type":"object","title":"Input"},"resolved":{"additionalProperties":true,"type":"object","title":"Resolved"},"times":{"additionalProperties":{"type":"string"},"type":"object","title":"Times"},"bazi":{"additionalProperties":true,"type":"object","title":"Bazi"}},"type":"object","required":["input","resolved","times","bazi"],"title":"BaziResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}